[test]: Fix test_mirror with correct number of ACL tables
After creation of the ACL table, check both the number of ACL tables
and the number of ACL table groups.

After removal of the ACL table, correct the number of remaining
ACL tables.

Right now, after the ACL table is removed, the ACL table groups are
not removed correspondingly. This should be fixed in the future.

Signed-off-by: Shu0T1an ChenG <shuche@microsoft.com>